+++ ReleaseNotes.md
@@ -1,3 +1,13 @@
+@1.0.0
+* A pure JavaScript library for using the Virtual Keyboard is ready to use.
+
+-----
+
+@1.0.0-beta.02
+* A pure JavaScript library for using the Virtual Keyboard. First Release.
+
+-----
+
 @1.0.0-beta.01
 * A pure JavaScript library for using the Virtual Keyboard. Coming Soon.